bluetooth: update bluetooth-power platform data
Update bluetooth power driver to support both device tree and
board specific platform data to power on bluetooth chipset.

Current driver supports only gpio data from the device node.
Now the driver first checks for device tree node and if it
does not exist then checks for the platform data from the
board file.

Updated the platform data for the driver to provide all the
required vregs and gpio data to power on the bluetooth
chipset.

Change-Id: Ie71115ad7173fc527e46b334a8b548762b40aa46
Signed-off-by: Ram Mohan Korukonda <rkorukon@codeaurora.org>