thermal: Possible fix for random reboots
It was recently brought to my attention that there were random reboots
reported in ivanich's CM thread and the cause for these were thought to
be the in-kernel thermal driver. I've reviewed the code that was
included last August and compared it with showp1984's bricked-ville
kernel. There were a significant amount of changes from the hammerhead
version previously included. I've ported the ville code over to our
pyramid in the hopes that it will fix the random reboots. Most of this
commit is from the bricked-ville master branch. I take credit only for
the pyramid port and since I'm not going to waste my time breaking that
out, I'll just let showp have it all. :)

Note that this commit will break most other devices that are still in
the repository but since this is now named a _pyramid kernel, I'm not
going to spend the time porting the other devices (at least not at this
time).

Change-Id: I4e162dd663b4c53781128880eed6343e977537d0